requirement.txt
napari-threedee
magicgui
qtpy
ruamel.yaml
vtk
numpy
pathlib
libigl<2.5.0